+++ lib/src/autoimpl/for_deref.rs
