+++ lib/src/fields.rs
@@ -76,8 +76,8 @@ pub struct Field {
     /// Optional field initializer.
     ///
     /// This is considered legal input when parsing, but not legal output. An
-    /// attribute rule such as [`crate::AttrImplDefault`] must remove the
-    /// initializer before output is generated.
+    /// attribute rule such as [`AttrImplDefault`](crate::scope::AttrImplDefault)
+    /// must remove the initializer before output is generated.
     pub assign: Option<(Token![=], Expr)>,
 }
